Use boringSSL as external repository
gRPC depends on boringSSL as an external repo.

Due to other projects also importing boringSSL, it is unclear if this is
being used as a package instead of a external repo (considering the
existing BUILD file). Therefore we declare a new repo instead of adding
WORKSPACE directly to the package.

Test: bazel run @boringssl//:bssl -- speed
Bug: 242091931
Change-Id: I01f0000ea0b01367fae8ca2bc63ec5fbb83a03ef