Replace the cquery helper to output more info.
This replaces the helper script that prints public headers to one that
prints a json with:

1. Path to the archive file
2. A list of include directories
3. A list of exported defines

Note that the include directories are what was used to compile the
target itself, and thus usually a superset of what is actually needed to
consume the library.

Lists are printed as semicollon separated strings (cmake lists) for
easier consumption by cmake.

Test: bazel cquery --output=starlark \
      --starlark:file=build/bazel/utils/cmake.cquery.bzl \
      @com_github_cares_cares//:ares
Bug: 274180576
Change-Id: I27c8a917769be93b3fcd6de2b45c32d22b5fee84