Detect MSVC tools and create import target.
A "msvc_tools_repository" rule is added to detect installed VC tools.
The "tool_versions" attribute can be used to restrict the tool versions,
and allows an empty string to use the latest version found.

A toolchain import target is created to add the include and lib paths.
The target now takes an additional "lib_search_paths" attribute to add
lib paths without linking a lib.

Bug: 288495471
Change-Id: I4778df0bec37f4e11bc1301bfbfe083ca87972df