Avoid depending on the bazel-out symlink.
The "bazel-out" symlink is not created until "bazel build" is run.
Therefore it is not suitable for referring to e.g. include paths as
cmake requires them to exist at config time.

This change converts the paths to either relative to output_base or
workspace. To use the output, cmake scripts need to substitute
"${output_base}" and "${workspace}" variables.

Change-Id: Ib1f48501ae9897631594060526ab5b44b6473aa1