Make rust_proc_macro work on mac-arm64.
Procedure macros are rust compiler extensions and therefore must be built for the same platform as the compiler itself. We use a mac-x64 rust compiler via Rosetta on mac-arm64 (since the rust team do not provide a mac-arm64 binary). Therefore proc_macros (and their deps) must be built for x64, while everything else is built for arm64.

However, this info is unknown to bazel - and the bazel rule simply assumes that proc_macros are built for the execution platform. This change adds a drop-in wrapper for the proc_macro rule that hardcodes the platform to x64 when targeting macOS.

Bug: 339263032
Change-Id: Ibbfa8d082bc21f232249696498b103da0a433827