macOS: Link against libc++.dylib to fix crash in libavd.dylib
We encountered crashes on macOS when loading libavd.dylib due to
a symbol initialization issue with std::__1::__stdinbuf<char>::imbue.
The root cause was linking against the static version of libc++ (.a),
leading to an isolated global variable within libavd.dylib.

This change modifies the build to link against the dynamic version of
libc++ (.dylib), ensuring shared global variables and resolving the
crash.

Bug: 354118899
Change-Id: Idfbacf7cd6110fa38065b98e233160326c1ace29