Revert^2 "Use c++ headers from macOS SDK."
This reverts commit 34ebfcfddac78aeb3de0f779b61f8437aec48150.

Reason for revert: Have another attempt (go/abtd passes)

Original Commit Message:

Use c++ headers from macOS SDK.

With aosp/3184801, we are actually linking to the libc++.dylib from the
latest SDK in the currently selected Xcode (or CLT). This poses the
following possible issues:

1. We are still compiling with the hermetic c++ headers, which are
   generally newer than what comes with the SDK. So linking to the
   SDK libc++ can cause compatibility issues.
2. The toolchain can be configured to use an older SDK from the same
   Xcode, so always linking to the latest SDK can be the wrong
   choice.

Ideally we should link to the hermetic libc++.dylib, but that requires
bazel to issue the correct rpath, which is only achieved with
`--dynamic_mode=fully` - but we don't want that since most lib should
still be statically linked.

This change make it explicit that we are using libc++ from the SDK, and
we use the headers from the same SDK. We will have to do this until
bazel supports dynamic toolchain libs in static mode.

Change-Id: I5eeb98eee28ef657875a1d5132a96f324c686c1f