Make dsym aspect work with rust binaries.
When rust binaries already declare a "dsym_folder" output group, we simply pass it through. We also rename the output group produced by the aspect to "dsym_folder" as well to be consistent.

rules_rust is also patched to not declare the dsym folder when it shouldn't (using cc_common.link).

Bug: 331193952
Change-Id: I23120c756d489865f0f2bcd0064b5af05c23a4fa