Fix toolchain to work with rules_go
rules_go fails to call clang when the wrapper binary is used, so this is removed. b/409795620 tracks adding these features back to the macos cc toolchain.

For another undetermined reason, using -fwhole-program-vtables also causes rules_go to fail and complain that -flto is needed. This is strange since -flto is provided. Maybe this can be added back with more debugging.

Bug: 371633138
Change-Id: I40727dd93a0f4bdfc2da88673ef5952807a5e163