build(deps): Add qt_prebuilt module dependency
This commit introduces the `qt_prebuilt` Bazel module (version 6.5.3)
as a dependency in the `toplevel.MODULE.bazel` file.

This is a prerequisite for building Qt-based components within the
Bazel build system.

The `toplevel.MODULE.bazel.lock` file is also updated to reflect this
new dependency and its associated transitive dependencies.

Bug: 437364713
Change-Id: Ia2ff76e45b45a87d93fb059e77865bd39dfcabcb