Define toolchains with module extension.
A toolchain extension is added to generate all the toolchain related
repos. Since extension generated repos are not visible to the
goldfish_build module itself, all the toolchain definitions are moved to
a generated hub repo. They are registered by the goldfish_build module.

Also bump the module version due to changes in MODULE.bazel.

Bug: 432262765
Change-Id: Ie1c224f85376518b421b1f18f83ab1d0d039c1ed