[build] Make cc wrapper to work with golang
Golang calls our wrapper from a different location rather than the execroot. The wrapper also fails to correctly wait for the subprocess when used with go compiler.

This change tries to detect the go environment and make corretions to the paths. Subprocess handling is also improved to avoid premature returns.

Fixes: 460273853
Test: n/a
Change-Id: I36936dfa9e494c56a912c606d70918636a6a6964