Make libxcb a shared library dependency.
It's wrong to link in libxcb statically if other libraries link to it
dynamically, as X connections being passed around contain cached values
that will resolve to different things between the two.

Also, our libxcb library is not the same as a system's libxcb library;
we actually bundle all the extensions into one library. To ensure binaries
that take a dependency on this actually use the bundled version and not
acctidentally pick up the system's libxcb, we rename it to libxcb-aemu.so.

Bug: 428923963
Test: lavapipe build
Change-Id: I15c6fa480ba9b5aae77bee3b048c708ae4cdd334